feat: Refactor module to inheritance structure
