+++ src/__init__.py
@@ -1,4 +1,4 @@
-from .models.vae import VAE
-from .models.vanila_ae import AE
+from .modules.vae import VAE
+from .modules.vanila_ae import AE
 
 __all__ = ["AE", "VAE"]
